Explicit mamba install for bioconductor-onlinefdr
The BioConductor preferred way to install packages is within an R
session with BiocManager::Install(). While I could get this to work
inside of a running container, I couldn't get it to work in the
definition file. Discovered that there was a bioconda channel which had
it as a package mamba could install.